feat(v0.2.0): 计划任务双级开关 writer（T41b）
- writers/task.rs: ITaskService COM 写任务级 / 触发器级 Enabled
- 两级开关分别写，恢复只动被改的那一级（T6 硬规则 1）
- read_states 读取两级状态供乐观锁/预演/快照
- 入口 guard::check，Locked 任务直接拒绝

测试: cargo test 214 passed / clippy -D warnings 干净

+++ src-tauri/src/writers/approved.rs
@@ -85,7 +85,7 @@ fn write_record(machine: bool, sub: &str, name: &str, enabled: bool) -> Result<(
     bytes[0] = if enabled { 0x02 } else { 0x03 };
     let now = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs() as u64)
+        .map(|d| d.as_secs())
         .unwrap_or(0);
     // FILETIME 是 100ns 粒度，这里把秒放进低 4 字节（足够排序 + 人类可读）
     let ft = now.wrapping_mul(10_000_000); // now(s) → 100ns  ticks
